+++ examples/src/bin/onoff_light_bt.rs
@@ -38,7 +38,7 @@ use embassy_futures::select::{select, select4};
 
 use log::{info, warn};
 
-use rand::RngCore;
+use rand::Rng;
 
 use rs_matter::crypto::{default_crypto, Crypto};
 use rs_matter::dm::clusters::app::level_control::LevelControlHooks;
@@ -141,7 +141,7 @@ fn run<N: NetCtl + WifiDiag + NetChangeNotif>(
     matter.startup(&kv)?;
 
     // Create the crypto instance
-    let crypto = default_crypto(rand::thread_rng(), DAC_PRIVKEY);
+    let crypto = default_crypto(rand::rng(), DAC_PRIVKEY);
 
     let mut rand = crypto.rand()?;
 
@@ -273,7 +273,7 @@ const NODE: Node<'static> = Node {
 /// The Data Model handler + meta-data for our Matter device.
 /// The handler is the root endpoint 0 handler plus the on-off handler and its descriptor.
 fn data_model<'a, OH: OnOffHooks, LH: LevelControlHooks, T>(
-    mut rand: impl RngCore + Copy,
+    mut rand: impl Rng + Copy,
     on_off: &'a on_off::OnOffHandler<'a, OH, LH>,
     wifi_diag: &'a dyn WifiDiag,
     net_ctl: T,